pyproject.toml
[tool.poetry]
name = "awesome-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.0"
uvicorn = "^0.17.6"
SQLAlchemy = "^1.4.32"
psycopg2 = "^2.9.3"
alembic = "^1.7.7"
hashids = "^1.3.1"
bcrypt = "^3.2.0"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
